# wp-content/plugins/gym-core/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gym Core REST API",
    "description": "REST API for Haanpaa Martial Arts gym management. All endpoints use the gym/v1 namespace and return a standard envelope: { success: true, data: <payload>, meta?: { pagination? } }.",
    "version": "2.1.0",
    "contact": {
      "name": "Andrew Wikel"
    }
  },
  "servers": [
    {
      "url": "/wp-json/gym/v1",
      "description": "WordPress REST API (gym/v1 namespace)"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-WP-Nonce",
        "description": "WordPress REST nonce (cookie authentication)"
      },
      "applicationPassword": {
        "type": "http",
        "scheme": "basic",
        "description": "WordPress Application Password"
      }
    },
...
  },
  "paths": {
...
    "/schedule": {
      "get": {
        "summary": "Weekly schedule view",
        "tags": ["Schedule"],
        "parameters": [
          { "name": "location", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "week_of", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "program", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Weekly schedule grouped by day" }
        }
      }
    },
...
    "/promotions/eligible": {
      "get": {
        "summary": "List promotion-eligible members",
        "tags": ["Promotions"],
        "security": [{ "cookieAuth": [] }, { "applicationPassword": [] }],
        "parameters": [
          { "name": "program", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 10, "maximum": 100 } }
        ],
        "responses": {
          "200": { "description": "Paginated eligible members" },
          "403": { "description": "Requires manage_woocommerce" }
        }
      }
    },
    "/promotions/recommend": {
      "post": {
        "summary": "Set coach recommendation for promotion",
        "tags": ["Promotions"],
        "security": [{ "cookieAuth": [] }, { "applicationPassword": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["user_id", "program"],
                "properties": {
                  "user_id": { "type": "integer" },
                  "program": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Recommendation set" },
          "403": { "description": "Requires gym_promote_student" },
          "404": { "description": "Member not found" }
        }
      }
    },
...
    "/briefings/class/{class_id}": {
      "get": {
        "summary": "Generate briefing for a specific class",
        "tags": ["Briefings"],
        "security": [{ "cookieAuth": [] }, { "applicationPassword": [] }],
        "parameters": [
          { "name": "class_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Class briefing with roster, stats, announcements" },
          "403": { "description": "Requires gym_view_briefing or manage_woocommerce" }
        }
      }
    },
...
  },
...
}